Better explanations and extra functionality in `MutableBody` (#3382)
`MutableBody` is the core data structure for MIR manipulation in
instrumentation passes. However, its public methods have limited
documentation and unclear semantics. This slowed down the development of
instrumentation passes.

This PR aims to fix that by:
- Clarifying how source instruction shifts when the methods are called;
- Adding functionality for inserting basic blocks;
- Expanding the support for different terminator types.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.