Add missing padding needded for alignment (#3401)
This PR adds extra padding for enums when needed for alignment. The
missing alignment was exposed in #2857 where the layout size from MIR
for `der::error::ErrorKind` was 48 bytes, and from `codegen_enum` was
only 41 bytes. The difference turned out to be caused by an 8-byte
alignment that the compiler uses for this enum.

Resolves #2857 
Resolves #3318