Handle intrinsics systematically (#3422)
Since compiler intrinsics are handled both in codegen and
instrumentation, it would make sense to have a single source of truth
with regard to which intrinsics we support and update it systematically.

This PR introduces an `Intrinsic` enum, which allows to parse intrinsic
name into one of the predefined variants we support or a catch-all
`Unsupported` variant.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.