Update deny action (#3461)
The current `cargo deny` configuration in `deny.toml` uses several keys
that have been deprecated. This PR removes the deprecated keys, and
updates the deny action to use v2 (as well as renames it from
`audit.yml` to `deny.yml`).

The only semantic difference is that `cargo deny` will now reject crates
that are maintained or have a notice on them, whereas previously, our
configuration set both to "warn". As mentioned in the docs though, one
can add an "ignore" if needed to bypass those advisories:


https://embarkstudios.github.io/cargo-deny/checks/advisories/cfg.html#the-version-field-optional

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.