Clarify comment in RFC Template (#3462)
Looking at the RFC template file, I thought that the recommendation to
leave the software design section empty referred to everything below the
comment, rather than just that section. This PR updates the comment.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.