Extend name resolution to support qualified paths (Partial Fix) (#3457)
This is the first part needed for us to add support to stubbing trait
implementations (https://github.com/model-checking/kani/issues/1997) and
primitive types (https://github.com/model-checking/kani/issues/2646).

After this change, we need to change stubs to accept a `FnResolution`
instead of `FnDef`. We also need to find out how to retrieve methods of
primitive types.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>
Co-authored-by: Felipe R. Monteiro <rms.felipe@gmail.com>