RFC for List Subcommand (#3463)
RFC for the `kani list` subcommand.

I saw the comment [in the
template](https://github.com/model-checking/kani/blob/main/rfc/src/template.md#software-design)
to leave the Software Design section empty, but I was looking at the raw
MD file and made the mistake of thinking it referred to all sections
below the comment. In other words, I thought when it said "We recommend
you to leave this empty for the first version of your RFC" it meant that
everything below the comment should be empty (so Rationale, Open
Questions, and Future Work) and not Software Design. I realized my
mistake when I was making this PR, but I figured I may as well leave it
since I already wrote it. I [opened a
PR](https://github.com/model-checking/kani/pull/3462) to update the
comment.

Resolves [#2573](https://github.com/model-checking/kani/issues/2573),
[#1612](#https://github.com/model-checking/kani/issues/1612).

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.