Update Kani Book (#3474)
Add the following to the book:

- **Reference section on contracts**. The point of this section is not
to be an exhaustive explanation--we have good documentation on contracts
already, and we don't want to maintain it in two places. The goal of the
section is to give readers an intuition for the main idea and then
direct them to more detailed resources.
- **Add a reference to the blog**--it's a good resource to learn about
Kani and it would be nice to have an easier way to find it. Inspired by
[this
post](https://users.rust-lang.org/t/can-somebody-explain-how-kani-verifier-works/113918).
- **Move the crates documentation earlier.** This partly addresses
[#3029](https://github.com/model-checking/kani/issues/3029) by at least
making our crates documentation more visible in the book. (I don't have
this PR as resolving that issue, since we would need to actually publish
the docs on `crates.io` to do that.)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.