Replace `proc-macro-error` with `proc-macro-error2` (#3493)
The following security advisory has been issued today regarding the
`proc-macro-error` crate which is causing our CI to fail:

https://rustsec.org/advisories/RUSTSEC-2024-0370

Replacing the crate with `proc-macro-error2` which is a fork of that
crate.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.