Enable stubbing and function contracts for primitive types (#3496)
Previously, Kani was not able to resolve methods that belong to
primitive types, which would not allow users to specify primitive types
as stub target, neither for contracts.

Thus, extend the name resolution to be able to convert a type expression
into a `Ty` for primitive types, and add a new method to search for
function implementation for primitive types.

I also moved the type resolution logic to its own module to make it a
bit cleaner.

Resolves #2646 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.