Add experimental API to generate arbitrary pointers
Pointer generator that can be used to generate an arbitrary pointer.

This generator allows users to build pointers with different allocation
status, initialization and alignemnt.
It contains an internal buffer that it uses to generate `InBounds` and
`OutBounds` pointers.
In those cases, the pointers will have the same provenance
as the generator, and the same lifetime.