Append harness name to the graph file (debug feature only) (#3528)
Our debug feature that dumps the call graph into a file was using the
crate name, which meant that the call graph kept getting overriden.

Instead, append the name of the harness to the file.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.