Update to CBMC 6.3.1 and fix auto-update script (#3537)
Update to CBMC release 6.3.1. This release includes a full fix for the
build problem that we carried a patch for (in #3431 and #3436), thus
dropping the patching step.

The automatic update of CBMC requires actually installing that newer
version of CBMC, else regression tests fail as seen in
https://github.com/model-checking/kani/actions/runs/10987766383/job/30503118786.

Resolves: #3507, #3536

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.