Add experimental API to generate arbitrary pointers (#3538)
This change adds a pointer generator that can non-deterministically
generate a pointer with different properties.

This generator allows users to build pointers with different allocation
status, initialization and alignment.
It contains an internal buffer that it uses to generate `InBounds` and
`OutOfBounds` pointers.
In those cases, the pointers will have the same provenance as the
generator, and the same lifetime.

This approach is different than generating a pointer from an arbitrary
`usize`. Kani uses demonic non-determinism to track
allocation lifetimes, which makes hard to reason about during
verification. I.e., one cannot assume a pointer is valid, and
initialized, and this can only be accomplished by manually tracking the
pointer status.

I added the new API under `-Z mem-predicates` since it allows reasoning
about memory, and I was hoping this wouldn't need another unstable flag.
:smile:

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.

---------

Co-authored-by: Zyad Hassan <88045115+zhassan-aws@users.noreply.github.com>