Add an LLBC backend (#3514)
This PR adds a new codegen backend that generates low-level borrow
calculus (LLBC), which is the format defined by
[Charon](https://github.com/AeneasVerif/charon) and
[Aeneas](https://github.com/AeneasVerif/aeneas).

The backend can be invoked using `-Zaeneas`, and will generate a `.llbc`
file. This file can then be passed to Aeneas to generate Lean for
example. Currently, Aeneas needs to be manually run on the generated
LLBC.

The PR translates stable MIR to unstructured LLBC (ULLBC) and then uses
the Charon library to apply transformations that reconstruct the control
flow to regain some of the high-level structure (loops, if statements,
etc.).

In this PR, very few MIR constructs are handled, but the PR is already
pretty big. More support to come in subsequent PRs.

Call-outs:

- Charon is currently not released on crates.io, so the library is added
as a dependency through GitHub with a particular commit.
- The Kani driver is currenty heavily tailored towards CBMC, so
currently it's expecting a goto file, and thus errors out when the
Aeneas backend is invoked.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.