Fix the computation of the number of bytes of a pointer offset (#3584)
This PR fixes the logic that computes the number of bytes of a pointer
offset. The logic was incorrectly using the size of the pointer as
opposed to the size of the pointee.

This fixes the soundness issue in #3582 (but not the spurious failures).

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.