Remove the overflow checks for wrapping_offset (#3589)
This PR removes the overflow checks done for the `arith_offset`
intrinsic, which is used in implementation of `ptr::wrapping_offset`.
According to the documentation, this operation is always safe:


https://doc.rust-lang.org/std/primitive.pointer.html#method.wrapping_offset
https://doc.rust-lang.org/std/intrinsics/fn.arith_offset.html

See
https://github.com/model-checking/kani/issues/3582#issuecomment-2403453345
for the context.

Resolves #3582 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.