`kani-cov`: A coverage tool for Kani (#3121)
This PR introduces `kani-cov`, a coverage-oriented tool for Kani. This
new tool will help users aggregate raw coverage results produced by Kani
and generate human-readable coverage summaries and reports.

`kani-cov` allows using three different subcommands: `merge`, `summary`
and `report`. In the following, we will show how each one of these
commands can be used.

### The `merge` subcommand

The `merge` subcommand takes raw coverage results from Kani (AKA
"kaniraw" files) and generates aggregated coverage results which are
stored into another file (AKA a "kanicov" file). This new file can then
be used for generating summaries or reports.

For example, if we assume this Rust code is in `main.rs`:

```rust
1 fn _other_function() {
2    println!("Hello, world!");
3 }
4
5 fn test_cov(val: u32) -> bool {
6     if val < 3 || val == 42 {
7         true
8     } else {
9         false
10    }
11 }
12
13 #[cfg_attr(kani, kani::proof)]
14 fn main() {
15    let test1 = test_cov(1);
16    let test2 = test_cov(2);
17    assert!(test1);
18    assert!(test2);
19 }
```
and we execute Kani with coverage enabled
```
kani main.rs --coverage -Zsource-coverage
```
the raw coverage results will be saved to a folder
```
[info] Coverage results saved to /home/ubuntu/coverage-experiments/src/kanicov_2024-09-23_23-49
```
we now can aggregate those results with the `merge` subcommand
```
kani-cov merge kanicov_2024-09-23_23-49/*kaniraw.json
```
which by default produces a `default_kanicov.json` file. We can also use
the `--output <path>` option to change the name of the file.

The type of file we just produced is known as the "profile" or the
"kanicov" file, and contains aggregated coverage results which may have
been originated in multiple verification sessions. This file is a
required input for the other two subcommands.

### The `summary` subcommand

Now that we have both the "kanicov" file and the "kanimap" file, we are
ready to produce coverage metrics with the `summary` subcommand:

```
kani-cov summary kanicov_2024-09-23_23-49/kanicov_2024-09-23_23-49_kanimap.json --profile default_kanicov.json
```

This outputs a table including coverage metrics for functions, lines and
regions for each file referenced in the coverage mappings:

```
| Filename                                      | Function (%) | Line (%)     | Region (%)  |
| --------------------------------------------- | ------------ | ------------ | ----------- |
| /home/ubuntu/coverage-experiments/src/main.rs | 2/3 (66.67)  | 8/12 (66.67) | 4/6 (66.67) |
```

Since the default format is markdown, the table can also be rendered as
in this PR description:

| Filename | Function (%) | Line (%) | Region (%) |
| --------------------------------------------- | ------------ |
------------ | ----------- |
| /home/ubuntu/coverage-experiments/src/main.rs | 2/3 (66.67) | 8/12
(66.67) | 4/6 (66.67) |

At present, no other formats are available. But the JSON and CSV formats
would be really nice and easy to have.

### The `report` subcommand

Coverage reports go a little further and allow users to visualize
coverage information on the source code.

The `report` requires the same input files as the `summary` subcommand:

```
kani-cov report kanicov_2024-09-23_23-49/kanicov_2024-09-23_23-49_kanimap.json --profile default_kanicov.json
```

When calling this command from a terminal, it will render like this:

![Screenshot 2024-10-04 at 5 49 51
PM](https://github.com/user-attachments/assets/65c12d62-4dd8-40e6-bfe9-8fc0e70127d2)

However, if you're redirecting the output to a file, it will back off to
another text-based mode which simply uses ` ``` ` and ` ''' ` to
represent opening and closing escapes for the coverage regions. This is
just another format called `escapes` which can be explicitly requested
as

```
kani-cov report kanicov_2024-09-23_23-49/kanicov_2024-09-23_23-49_kanimap.json --profile default_kanicov.json --format=escapes
```

Similar to the `summary` command, this will iterate over the files in
the coverage mappings, compute coverage information and highlight any
uncovered regions (except #3543).

### Testing

This PR also changes `compiletest` to automatically call the `merge` and
`report` subcommands on the `coverage` test suite. The tests have been
blessed using the `--fix-expected` option and this Python3 script that
removes the first line (path to the file) and adds `\` after each line
so an exact match is expected.

```
import os

def process_files(directory):
    for filename in os.listdir(directory):
        if filename == "expected":
            file_path = os.path.join(directory, filename)
            with open(file_path, "r") as file:
                lines = file.readlines()
            
            # Remove the first line
            lines = lines[1:]
            
            # Add the character `\` after every line
            processed_lines = []
            for i, line in enumerate(lines):
                processed_lines.append(line.replace("\n", "\\\n"))
            
            # Write the processed lines back to the file
            with open(file_path, "w") as file:
                file.writelines(processed_lines)
```

### Callouts / Known issues

This PR only introduces `kani-cov`. It doesn't plug `kani-cov` into
`kani-driver` so it's automatically run when using the `--coverage`
option, nor it prepares the `kani-cov` binary for distribution. This is
better to address later.

This PR also spawns a few issues:
- #3543 forces us to filter out certain coverage results to avoid issues
with the report command. It is important that we investigate where these
coverage results come from and their semantics as this represents a
soundness issue.
- #3541 is to avoid duplicating certain data structures that currently
live in `kani-driver`. We should probably have a different crate for
these structures and import them from `kani-driver` and `kani-cov`, but
this is more of a distribution problem that doesn't need to be tackled
here.
- #3542 requires us to match filenames when extracting function coverage
results. It's not a big deal, but it'll likely improve the tool's
robustness in the long term.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.