Exclude Charon from workspace (#3580)
@celinval pointed out that `cargo update` will attempt to update the
`charon` submodule as well.

Exclude the `charon` submodule from the workspace so that it's not
considered when running `cargo update`, `cargo deny`, `cargo clippy`,
and `cargo fmt`. The PR also reverts some of the configuration changes
made in #3514.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.