Implement proper function pointer handling for validity checks (#3606)
We previously had a `todo!()` for the validity check handling of
function pointers. However, the validity check only cares about
intrinsics calls, which can only be done directly. Thus, remove the
`todo!()` and refactor the code so that it's clear why we don't care
about that case.

I added a test that I was working on when I bumped into this issue.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.