Add fn that checks pointers point to same allocation (#3583)
This changes adds a new function that returns whether two pointers
points to the same allocated object. Since we cannot reason about
addresses to objects that are dead or haven't been allocated, we panic
if the user provides dangling pointers.

## Call-out:

- Creating this as draft for now for early feedback

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.